pyproject.toml
[project]
name = "horatio"
version = "0.1.0"
description = ""
authors = [{ name = "Jeffrey Yun", email = "thisishaoyu@gmail.com" }]
dependencies = [
    "python>=3.11,<4",
    "fastapi>=0.115.0,<0.116.0",
    "uvicorn>=0.31.1,<0.32.0",
    "openai>=1.51.2,<1.52.0",
    "python-dotenv>=1.0.1,<1.1.0",
    "hypercorn>=0.17.3,<0.18.0",
]

[tool.poetry]
name = "horatio"
version = "0.1.0"
description = ""
authors = ["Jeffrey Yun <thisishaoyu@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.31.1"
openai = "^1.51.2"
python-dotenv = "^1.0.1"
hypercorn = "^0.17.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
